+++ controllers/index.js
@@ -1,0 +1,21 @@
+const {getListContacts,
+    getContact,
+    addNewContact,
+    deleteContact,
+    changeContact } = require('./controllers')
+    
+
+
+
+
+
+
+
+
+module.exports = {
+    getListContacts,
+    getContact,
+    addNewContact,
+    deleteContact,
+    changeContact
+}
