+++ schemas/index.js
@@ -1,0 +1,9 @@
+
+const addSchema = require('./addSchema');
+
+
+
+
+module.exports = {
+    addSchema
+}
